requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Testing==0.8.1
factory-boy==3.2.1
behave==1.2.6
selenium==4.10.0
pytest==7.4.0
requests==2.31.0
Werkzeug==2.3.6
compare==0.2b0